requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
pandas
plotly